// src/modules/attendance/enums/attendance-status.enum.ts
export enum AttendanceStatus {
  PRESENT = 'PRESENT',
  HALF_DAY = 'HALF_DAY',
  ABSENT = 'ABSENT',
  LATE = 'LATE',
  ON_LEAVE = 'ON_LEAVE',
  HOLIDAY = 'HOLIDAY',
  WEEKLY_OFF = 'WEEKLY_OFF',
  MISSED_CHECKOUT = 'MISSED_CHECKOUT',
  OVERTIME = 'OVERTIME',
  PENDING_APPROVAL = 'PENDING_APPROVAL',
}